framework/st_things : Network related module merged into things_network
- Refactoring of things_network
- Move wifi_manager related logic
- Remove unnecessary logic
