+++ framework/src/st_things/things_stack/things_api.h
@@ -108,7 +108,9 @@ int things_get_attributes_by_resource_type(const char *res_type, int *count, thi
 
 bool things_is_collection_resource(const char *res_uri);
 
-int things_get_child_resources(const char *col_res_uri, int *count, things_resource_info_s ***child_resources);
+int things_get_child_resources(const char *col_res_uri, int *count, things_resource_info_s ***child_resurces);
+
+int things_is_things_module_inited(void);
 
 #ifdef __cplusplus
 }
